+++ nflow-engine/src/main/java/io/nflow/engine/internal/dao/WorkflowInstanceDao.java
@@ -67,9 +67,7 @@ import org.springframework.jdbc.core.support.AbstractInterruptibleBatchPreparedS
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.TransactionCallbackWithoutResult;
 import org.springframework.transaction.support.TransactionTemplate;
 import org.springframework.util.Assert;
 
@@ -94,8 +92,8 @@ import jakarta.inject.Singleton;
  * Use setter injection because constructor injection may not work when nFlow is used in some legacy systems.
  */
 @Component
-@SuppressFBWarnings(value = { "SIC_INNER_SHOULD_BE_STATIC_ANON", "NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE" },
-    justification = "common jdbctemplate practice, npe is unlikely")
+@SuppressFBWarnings(value = "SIC_INNER_SHOULD_BE_STATIC_ANON",
+    justification = "common jdbctemplate practice")
 @Singleton
 public class WorkflowInstanceDao {
 
@@ -376,23 +374,20 @@ public class WorkflowInstanceDao {
   private void updateWorkflowInstanceWithTransaction(final WorkflowInstance instance, final WorkflowInstanceAction action,
       final List<WorkflowInstance> childWorkflows, final List<WorkflowInstance> workflows,
       final Map<String, String> changedStateVariables) {
-    transaction.execute(new TransactionCallbackWithoutResult() {
-      @Override
-      protected void doInTransactionWithoutResult(TransactionStatus status) {
-        int updated = updateWorkflowInstance(instance);
-        if (updated == 0) {
-          return;
-        }
-        long parentActionId = insertWorkflowInstanceAction(action);
-        insertVariables(action.workflowInstanceId, parentActionId, changedStateVariables);
-        for (WorkflowInstance childTemplate : childWorkflows) {
-          WorkflowInstance childWorkflow = new WorkflowInstance.Builder(childTemplate).setParentWorkflowId(instance.id)
-              .setParentActionId(parentActionId).build();
-          insertWorkflowInstance(childWorkflow);
-        }
-        for (WorkflowInstance workflow : workflows) {
-          insertWorkflowInstance(workflow);
-        }
+    transaction.executeWithoutResult(status -> {
+      int updated = updateWorkflowInstance(instance);
+      if (updated == 0) {
+        return;
+      }
+      long parentActionId = insertWorkflowInstanceAction(action);
+      insertVariables(action.workflowInstanceId, parentActionId, changedStateVariables);
+      for (WorkflowInstance childTemplate : childWorkflows) {
+        WorkflowInstance childWorkflow = new WorkflowInstance.Builder(childTemplate).setParentWorkflowId(instance.id)
+            .setParentActionId(parentActionId).build();
+        insertWorkflowInstance(childWorkflow);
+      }
+      for (WorkflowInstance workflow : workflows) {
+        insertWorkflowInstance(workflow);
       }
     });
   }