+++ nflow-engine/src/test/java/io/nflow/engine/guice/EngineModuleTest.java
@@ -7,7 +7,6 @@ import static org.hamcrest.Matchers.isA;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 
-import java.io.IOException;
 import java.util.Properties;
 import java.util.concurrent.ThreadFactory;
 
@@ -22,14 +21,13 @@ import org.springframework.scheduling.concurrent.CustomizableThreadFactory;
 import org.springframework.transaction.support.TransactionTemplate;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Key;
 import com.zaxxer.hikari.HikariDataSource;
 
 import io.nflow.engine.config.EngineConfiguration;
-import io.nflow.engine.config.EngineConfiguration.EngineObjectMapperSupplier;
+import io.nflow.engine.config.EngineConfiguration.EngineJsonMapperSupplier;
 import io.nflow.engine.config.NFlow;
 import io.nflow.engine.config.db.H2DatabaseConfiguration;
 import io.nflow.engine.internal.executor.WorkflowInstanceExecutor;
@@ -42,11 +40,12 @@ import io.nflow.engine.service.StatisticsService;
 import io.nflow.engine.service.WorkflowDefinitionService;
 import io.nflow.engine.service.WorkflowExecutorService;
 import io.nflow.engine.service.WorkflowInstanceService;
+import tools.jackson.databind.json.JsonMapper;
 
 public class EngineModuleTest {
 
   @Test
-  public void testEngineConfiguration() throws IOException {
+  public void testEngineConfiguration() {
     Properties props = new Properties();
     props.setProperty("nflow.db.type", "h2");
     props.setProperty("nflow.executor.thread.count", "1");
@@ -65,10 +64,10 @@ public class EngineModuleTest {
     assertThat(((CustomizableThreadFactory) factory).getThreadNamePrefix(), is("nflow-executor-"));
     assertThat(((CustomizableThreadFactory) factory).getThreadGroup().getName(), is("nflow"));
 
-    ObjectMapper mapper = injector.getInstance(Key.get(EngineObjectMapperSupplier.class, NFlow.class)).get();
+    JsonMapper mapper = injector.getInstance(Key.get(EngineJsonMapperSupplier.class, NFlow.class)).get();
     String nowS = mapper.writeValueAsString(DateTime.now());
-    assertThat(mapper.readerFor(DateTime.class).readValue(nowS, DateTime.class), isA(DateTime.class));
-    assertThat(mapper.getSerializationConfig().getDefaultPropertyInclusion().getValueInclusion(),
+    assertThat(mapper.readerFor(DateTime.class).<DateTime>readValue(nowS), isA(DateTime.class));
+    assertThat(/* TODO serializationConfig() is not to be used by application code in Jackson 3 (see https://github.com/FasterXML/jackson-databind/blob/3.x/src/main/java/tools/jackson/databind/ObjectMapper.java#L417). Consider using builder configuration instead. */mapper.serializationConfig().getDefaultPropertyInclusion().getValueInclusion(),
         is(JsonInclude.Include.NON_EMPTY));
 
     DataSource dataSource = injector.getInstance(Key.get(DataSource.class, NFlow.class));