+++ nflow-tests/src/test/java/io/nflow/tests/CreditApplicationWorkflowTest.java
@@ -1,9 +1,9 @@
 package io.nflow.tests;
 
 import static io.nflow.engine.workflow.instance.WorkflowInstanceAction.WorkflowActionType.stateExecution;
+import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON_TYPE;
 import static java.time.Duration.ofSeconds;
 import static java.util.Arrays.asList;
-import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON_TYPE;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
@@ -13,14 +13,10 @@ import static org.joda.time.DateTime.now;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-import jakarta.ws.rs.core.Response;
-
 import org.junit.jupiter.api.MethodOrderer;
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.nflow.rest.v1.msg.Action;
 import io.nflow.rest.v1.msg.CreateWorkflowInstanceRequest;
@@ -29,6 +25,8 @@ import io.nflow.rest.v1.msg.ErrorResponse;
 import io.nflow.rest.v1.msg.UpdateWorkflowInstanceRequest;
 import io.nflow.tests.demo.workflow.CreditApplicationWorkflow;
 import io.nflow.tests.extension.NflowServerConfig;
+import jakarta.ws.rs.core.Response;
+import tools.jackson.databind.json.JsonMapper;
 
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 public class CreditApplicationWorkflowTest extends AbstractNflowTest {
@@ -48,7 +46,7 @@ public class CreditApplicationWorkflowTest extends AbstractNflowTest {
     req = new CreateWorkflowInstanceRequest();
     req.type = "creditApplicationProcess";
     req.businessKey = UUID.randomUUID().toString();
-    req.stateVariables.put("requestData", (new ObjectMapper()).valueToTree(
+    req.stateVariables.put("requestData", (new JsonMapper()).valueToTree(
             new CreditApplicationWorkflow.CreditApplication("CUST123", new BigDecimal(100l))));
     req.externalId = UUID.randomUUID().toString();
     resp = createWorkflowInstance(req);